pyproject.toml
[project]
name = "meeting-memory-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain (>=1.2.6,<2.0.0)",
    "langchain-aws (>=1.2.1,<2.0.0)",
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
